+++ src/js/containers/agencySubmissionStats/AgencyDetailsContainer.jsx
@@ -9,7 +9,7 @@ import { Table, Pagination } from 'data-transparency-ui';
 import { throttle, isNull } from 'lodash';
 
 import { fetchAgency } from 'apis/agencyReporting';
-import BaseReportingPeriodRow from 'models/v2/aboutTheData/BaseReportingPeriodRow';
+import BaseReportingPeriodRow from 'models/v2/agencySubmissionStats/BaseReportingPeriodRow';
 import AgencyDownloadLinkCell from 'components/agencySubmissionStats/AgencyDownloadLinkCell';
 import CellWithModal from 'components/agencySubmissionStats/CellWithModal';
 import { agencyDetailsColumns } from './AgencyTableMapping';
